webpack: forward devServer config to webpack-dev-server
