+++ src/scripts/webpack/build-ts.js
@@ -2,10 +2,14 @@ import webpackBuild from './webpack.build.js';
 import compiler from './compiler';
 import { logger } from '../../util';
 import { statsOutputConfiguration } from './util';
+import getWebpackConfig from './getWebpackConfig';
 
-const getBuildConfig = (options) => webpackBuild({ ...options, jsLoader: 'typescript' });
+const webpackBuildObject = getWebpackConfig(
+  options => webpackBuild({ ...options, jsLoader: 'typescript' }),
+  'build-ts'
+);
 
-compiler(getBuildConfig, 'build-ts').run((err, stats) => {
+compiler(webpackBuildObject).run((err, stats) => {
   if (err) { throw err; }
   logger.webpack(stats.toString(statsOutputConfiguration));
 });
