+++ src/scripts/webpack/build.js
@@ -2,8 +2,11 @@ import webpackBuild from './webpack.build.js';
 import compiler from './compiler';
 import { logger } from '../../util';
 import { statsOutputConfiguration } from './util';
+import getWebpackConfig from './getWebpackConfig';
 
-compiler(webpackBuild, 'build').run((err, stats) => {
+const webpackBuildObject = getWebpackConfig(webpackBuild, 'build');
+
+compiler(webpackBuildObject).run((err, stats) => {
   if (err) { throw err; }
   logger.webpack(stats.toString(statsOutputConfiguration));
 });
