+++ src/scripts/webpack/dev-ts.js
@@ -3,20 +3,15 @@ import webpackServer from 'webpack-dev-server';
 import webpackConfig from './webpack.config.js';
 import compiler from './compiler';
 import getConfig from './config';
-import getPaths from './paths';
-import { statsOutputConfiguration } from './util';
-
-const getWebpackConfig = (options) => webpackConfig({ ...options, jsLoader: 'typescript' });
+import getWebpackConfig from './getWebpackConfig';
 
 const args = minimist(process.argv.slice(2));
-const paths = getPaths(args);
 
-const server = new webpackServer(compiler(getWebpackConfig, 'dev-ts'), {
-  contentBase: paths.BUILD,
-  hot: true,
-  stats: statsOutputConfiguration,
-  disableHostCheck: true,
-  historyApiFallback: { verbose: true }
-});
+const webpackConfigObject = getWebpackConfig(
+  options => webpackConfig({ ...options, jsLoader: 'typescript' }),
+  'dev-ts'
+);
+
+const server = new webpackServer(compiler(webpackConfigObject), webpackConfigObject.devServer);
 
 server.listen(getConfig(args).port);
